changed paging type to qint64
Changed in all paging related methods like chunk(), take(), ...

 - updated unit tests

+++ include/orm/query/concerns/buildsqueries.hpp
@@ -34,23 +34,24 @@ namespace Orm::Query::Concerns
         BuildsQueries &operator=(BuildsQueries &&) = delete;
 
         /*! Chunk the results of the query. */
-        bool chunk(int count,
-                   const std::function<bool(SqlQuery &results, int page)> &callback);
+        bool chunk(qint64 count,
+                   const std::function<bool(SqlQuery &results, qint64 page)> &callback);
         /*! Execute a callback over each item while chunking. */
-        bool each(const std::function<bool(SqlQuery &row, int index)> &callback,
-                  int count = 1000);
+        bool each(const std::function<bool(SqlQuery &row, qint64 index)> &callback,
+                  qint64 count = 1000);
 
         /*! Run a map over each item while chunking. */
 //        QVector<QSqlQuery>
-//        chunkMap(const std::function<void(QSqlQuery &row)> &callback, int count = 1000);
+//        chunkMap(const std::function<void(QSqlQuery &row)> &callback, qint64 count = 1000);
 
         /*! Chunk the results of a query by comparing IDs. */
-        bool chunkById(int count,
-                       const std::function<bool(SqlQuery &results, int page)> &callback,
+        bool chunkById(qint64 count,
+                       const std::function<
+                           bool(SqlQuery &results, qint64 page)> &callback,
                        const QString &column = "", const QString &alias = "");
         /*! Execute a callback over each item while chunking by ID. */
-        bool eachById(const std::function<bool(SqlQuery &row, int index)> &callback,
-                      int count = 1000, const QString &column = "",
+        bool eachById(const std::function<bool(SqlQuery &row, qint64 index)> &callback,
+                      qint64 count = 1000, const QString &column = "",
                       const QString &alias = "");
